webapi/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   starlette
    #   watchfiles
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via -r requirements.in
fastapi==0.110.2
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   requests
jinja2==3.1.4
    # via -r requirements.in
markupsafe==2.1.5
    # via jinja2
psycopg2-binary==2.9.9
    # via -r requirements.in
pydantic==2.7.1
    # via fastapi
pydantic-core==2.18.2
    # via pydantic
pyjwt==2.8.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via uvicorn
python-json-logger==2.0.7
    # via -r requirements.in
pytz==2024.1
    # via -r requirements.in
pyyaml==6.0.1
    # via uvicorn
requests==2.31.0
    # via -r requirements.in
sentry-sdk==2.1.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.37.2
    # via fastapi
typing-extensions==4.11.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.29.0
    # via -r requirements.in
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
